# openapi.yaml
openapi: 3.1.0
info:
  title: Vivant Valley Hosted AI API
  version: 0.1.0-contract
  license:
    name: Proprietary contract draft
    identifier: LicenseRef-Proprietary
  description: |
    Phase-1 contract for the hosted Vivant Valley service. LiteLLM and all
    provider credentials are private implementation details. Monetary values
    are integer credit_micros; request and response bodies must not contain
    secrets.
servers:
  - url: https://api.example.invalid
    description: Replace with the production HTTPS origin
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      operationId: register
      security: []
      summary: Register a web account
      description: Invitation code is required when registration_invite_required is enabled.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account and authenticated session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/ConflictError' }
        '429': { $ref: '#/components/responses/RateLimitError' }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Create a web session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated session; sets a Secure HttpOnly cookie
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401': { $ref: '#/components/responses/AuthError' }
        '429': { $ref: '#/components/responses/RateLimitError' }
...
  /api/v1/keys:
    get:
      tags: [Device keys]
      operationId: listDeviceKeys
      summary: List device keys
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Keys without plaintext secrets
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/DeviceKey' }
        '401': { $ref: '#/components/responses/AuthError' }
    post:
      tags: [Device keys]
      operationId: createDeviceKey
      summary: Create a device key
      security:
        - SessionCookie: []
      description: Plaintext key is returned exactly once and is never retrievable later.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeviceKeyRequest'
      responses:
        '201':
          description: Created key; plaintext is present only in this response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedDeviceKey'
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '429': { $ref: '#/components/responses/RateLimitError' }
...
  /v1/models:
    get:
      tags: [OpenAI compatibility]
      operationId: listModels
      summary: List public model aliases
      security:
        - DeviceBearer: []
      responses:
        '200':
          description: Public allow-listed model aliases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelList'
        '401': { $ref: '#/components/responses/DeviceAuthError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
  /v1/chat/completions:
    post:
      tags: [OpenAI compatibility]
      operationId: createChatCompletion
      summary: Create a chat completion
      security:
        - DeviceBearer: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
            examples:
              ordinary:
                value:
                  model: vv-dialogue
                  messages:
                    - role: system
                      content: You are an NPC.
                    - role: user
                      content: Hello!
                  max_tokens: 512
                  stream: false
              tool_round:
                value:
                  model: vv-dialogue
                  messages:
                    - role: system
                      content: Choose a permitted action.
                    - role: user
                      content: I would like to travel.
                  tools:
                    - type: function
                      function:
                        name: move_to
                        description: Start an agreed journey.
                        parameters:
                          type: object
                          properties:
                            destination_key: { type: string }
                          required: [destination_key]
                          additionalProperties: false
                  tool_choice: auto
                  stream: false
      responses:
        '200':
          description: Chat completion or an SSE stream when stream=true
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
            X-Balance-Micros:
              schema: { type: integer, format: int64 }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatCompletion' }
            text/event-stream:
              schema:
                type: string
                description: 'data: JSON chunks followed by data: [DONE]'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/DeviceAuthError' }
        '402':
          description: Balance or spend limit is insufficient
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '408': { $ref: '#/components/responses/TimeoutError' }
        '409': { $ref: '#/components/responses/ConflictError' }
        '413':
          description: Request body is too large
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '429': { $ref: '#/components/responses/RateLimitError' }
        '502': { $ref: '#/components/responses/UpstreamError' }
        '503': { $ref: '#/components/responses/UpstreamError' }
...
  /api/v1/admin/users/{userId}/credits:
    post:
      tags: [Admin]
      operationId: adminTopUp
      summary: Add manual credits to a user
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TopUpRequest' }
      responses:
        '200':
          description: New wallet balance
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Wallet' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { $ref: '#/components/responses/ConflictError' }
...
  /api/v1/admin/providers:
    get:
      tags: [Admin]
      operationId: adminListProviders
      summary: List managed upstream providers without secrets
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Provider metadata and masked key hints
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/AdminProvider' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
    post:
      tags: [Admin]
      operationId: adminCreateProvider
      summary: Add an encrypted upstream API credential
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProviderRequest' }
      responses:
        '201':
          description: Created provider; the API key is not returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminProvider' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
  /api/v1/admin/providers/{id}:
    patch:
      tags: [Admin]
      operationId: adminUpdateProvider
      summary: Update provider metadata or rotate its API key
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProviderRequest' }
      responses:
        '200':
          description: Updated provider metadata
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminProvider' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    delete:
      tags: [Admin]
      operationId: adminDeleteProvider
      summary: Delete an unused managed provider
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204': { description: Provider deleted }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { $ref: '#/components/responses/ConflictError' }
  /api/v1/admin/providers/{id}/test:
    post:
      tags: [Admin]
      operationId: adminTestProvider
      summary: Probe upstream Chat Completions and forced tool-call compatibility
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Provider returned a valid forced tool call; this probe may consume a small number of upstream tokens
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '502': { $ref: '#/components/responses/UpstreamError' }
...
components:
  securitySchemes:
    DeviceBearer:
      type: http
      scheme: bearer
      bearerFormat: vv_live_device_key
      description: Plaintext vv_live_ device key created in the web console.
    ModBearer:
      type: http
      scheme: bearer
      bearerFormat: vv_mod_session
      description: Opaque hosted-account session token issued to the Mod; never an upstream provider key.
    SessionCookie:
      type: apiKey
      in: cookie
      name: vv_session
      description: Secure HttpOnly web session cookie; mutating calls also need X-CSRF-Token.
    AdminGateCookie:
      type: apiKey
      in: cookie
      name: vv_admin_gate
      description: HttpOnly cookie issued after ADMIN_PAGE_PASSWORD verification; required with an administrator session.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: 1..128 character client-generated key; reused requests return the original result.
      schema: { type: string, minLength: 1, maxLength: 128 }
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Client correlation ID; server generates a UUID when absent.
      schema: { type: string, maxLength: 128 }
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 256 }
...
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    ValidationError:
      description: Invalid JSON, field, tool schema, alias or limit
      content:
        application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } }
    AuthError:
      description: Missing or invalid web session
      content:
        application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } }
    DeviceAuthError:
      description: Missing, revoked or expired device key
      content:
        application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } }
    ForbiddenError:
      description: Suspended account, disabled key or insufficient role
      content:
        application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } }
    NotFoundError:
      description: Resource not found
      content:
        application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } }
    ConflictError:
      description: Duplicate email, idempotency conflict or already consumed code
      content:
        application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } }
    RateLimitError:
      description: RPM or concurrent-request limit exceeded
      headers:
        Retry-After: { schema: { type: integer, minimum: 1 } }
      content:
        application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } }
    TimeoutError:
      description: Provider deadline exceeded; reservation released
      content:
        application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } }
    UpstreamError:
      description: Internal LiteLLM/provider failure; reservation released or pending reconciliation
      content:
        application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type, code]
          properties:
            message: { type: string }
            type: { type: string, enum: [validation_error, auth_error, billing_error, rate_limit_error, upstream_error, conflict_error, forbidden_error, timeout_error] }
            code: { type: string, examples: [insufficient_balance, invalid_device_key, stream_not_supported] }
            param: { type: [string, 'null'] }
            request_id: { type: string, format: uuid }
          additionalProperties: false
      additionalProperties: false
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 12, maxLength: 256, writeOnly: true }
        invitation_code: { type: string, minLength: 8, maxLength: 128, writeOnly: true }
      additionalProperties: false
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, maxLength: 256, writeOnly: true }
      additionalProperties: false
    AuthSession:
      type: object
      required: [user, csrf_token]
      properties:
        user: { $ref: '#/components/schemas/User' }
        csrf_token: { type: string, readOnly: true }
...
    User:
      type: object
      required: [id, email, role, status, created_at, wallet]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { type: string, enum: [user, admin] }
        status: { type: string, enum: [active, suspended, deleted] }
        created_at: { type: string, format: date-time }
        last_login_at: { type: [string, 'null'], format: date-time }
        wallet: { $ref: '#/components/schemas/Wallet' }
    Wallet:
      type: object
      required: [available_micros, reserved_micros, currency]
      properties:
        available_micros: { type: integer, format: int64, minimum: 0 }
        reserved_micros: { type: integer, format: int64, minimum: 0 }
        currency: { type: string, example: credit_micros }
    CreateDeviceKeyRequest:
      type: object
      properties:
        label: { type: string, maxLength: 80 }
        expires_at: { type: string, format: date-time }
      additionalProperties: false
    DeviceKey:
      type: object
      required: [id, display_prefix, label, created_at, revoked]
      properties:
        id: { type: string, format: uuid }
        display_prefix: { type: string, pattern: '^vv_live_[A-Za-z0-9_-]{4,}$' }
        label: { type: string }
        created_at: { type: string, format: date-time }
        last_used_at: { type: [string, 'null'], format: date-time }
        expires_at: { type: [string, 'null'], format: date-time }
        revoked: { type: boolean }
    CreatedDeviceKey:
      allOf:
        - $ref: '#/components/schemas/DeviceKey'
        - type: object
          required: [key]
          properties:
            key: { type: string, pattern: '^vv_live_[A-Za-z0-9_-]{32,}$', writeOnly: true }
    Model:
      type: object
      required: [id, object, owned_by]
      properties:
        id: { type: string, pattern: '^vv-[a-z0-9-]+$' }
        object: { type: string, enum: [model] }
        owned_by: { type: string, example: vivantvalley }
        permission: { type: array, items: { type: string } }
    ModelList:
      type: object
      required: [object, data]
      properties:
        object: { type: string, enum: [list] }
        data: { type: array, items: { $ref: '#/components/schemas/Model' } }
    ChatMessage:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [system, user, assistant, tool] }
        content:
          oneOf:
            - type: string
            - type: 'null'
            - type: array
              items:
                type: object
                required: [type]
                properties:
                  type: { type: string }
                  text: { type: string }
                additionalProperties: true
        name: { type: string, maxLength: 128 }
        tool_call_id: { type: string, maxLength: 256 }
        tool_calls:
          type: array
          items: { $ref: '#/components/schemas/ToolCall' }
      additionalProperties: false
...
    ToolDefinition:
      type: object
      required: [type, function]
      properties:
        type: { type: string, enum: [function] }
        function:
          type: object
          required: [name, parameters]
          properties:
            name: { type: string, maxLength: 128 }
            description: { type: string, maxLength: 4000 }
            parameters: { type: object, additionalProperties: true }
          additionalProperties: false
      additionalProperties: false
    ToolChoice:
      oneOf:
        - type: string
          enum: [none, auto, required]
        - type: object
          required: [type, function]
          properties:
            type: { type: string, enum: [function] }
            function:
              type: object
              required: [name]
              properties: { name: { type: string, maxLength: 128 } }
              additionalProperties: false
          additionalProperties: false
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model: { type: string, pattern: '^vv-[a-z0-9-]+$', maxLength: 64 }
        messages:
          type: array
          minItems: 1
          maxItems: 128
          items: { $ref: '#/components/schemas/ChatMessage' }
        stream: { type: boolean, default: false }
        tools:
          type: array
          maxItems: 32
          items: { $ref: '#/components/schemas/ToolDefinition' }
        tool_choice: { $ref: '#/components/schemas/ToolChoice' }
        max_tokens: { type: integer, minimum: 1, maximum: 8192 }
        max_completion_tokens: { type: integer, minimum: 1, maximum: 8192 }
        temperature: { type: number, minimum: 0, maximum: 2 }
        top_p: { type: number, minimum: 0, maximum: 1 }
        thinking:
          type: object
          description: Accepted for client compatibility; hosted routing forces non-thinking mode upstream.
          properties: { type: { type: string, enum: [enabled, disabled] } }
          additionalProperties: false
        reasoning_effort:
          type: string
          enum: [low, medium, high]
          description: Accepted for client compatibility but ignored while hosted routing is in non-thinking mode.
        enable_thinking:
          type: boolean
          description: Accepted for client compatibility; hosted routing forces this behavior off upstream.
        user: { type: string, maxLength: 128 }
      additionalProperties: false
    ChatCompletion:
      type: object
      required: [id, object, created, model, choices, usage]
      properties:
        id: { type: string }
        object: { type: string, enum: [chat.completion] }
        created: { type: integer, format: int64 }
        model: { type: string, pattern: '^vv-[a-z0-9-]+$' }
        choices:
          type: array
          minItems: 1
          items:
            type: object
            required: [index, message, finish_reason]
            properties:
              index: { type: integer, minimum: 0 }
              message: { $ref: '#/components/schemas/ChatMessage' }
              finish_reason: { type: [string, 'null'], enum: [stop, length, tool_calls, content_filter, 'null'] }
        usage: { $ref: '#/components/schemas/Usage' }
    Usage:
      type: object
      required: [prompt_tokens, completion_tokens, total_tokens]
      properties:
        prompt_tokens: { type: integer, minimum: 0 }
        completion_tokens: { type: integer, minimum: 0 }
        total_tokens: { type: integer, minimum: 0 }
        prompt_tokens_details:
          type: object
          properties: { cached_tokens: { type: integer, minimum: 0 } }
          additionalProperties: false
...
    TopUpRequest:
      type: object
      required: [amount_micros, reason]
      properties:
        amount_micros: { type: integer, format: int64, minimum: 1, maximum: 1000000000000 }
        reason: { type: string, minLength: 1, maxLength: 500 }
      additionalProperties: false
...
    AdminProvider:
      type: object
      required: [id, name, baseUrl, defaultModel, enabled, keyHint, source]
      properties:
        id: { type: [string, 'null'], format: uuid }
        name: { type: string }
        baseUrl: { type: string, format: uri }
        defaultModel: { type: string }
        enabled: { type: boolean }
        keyHint: { type: string, description: Masked key prefix/suffix only }
        createdAt: { type: [string, 'null'], format: date-time }
        updatedAt: { type: [string, 'null'], format: date-time }
        lastCheckedAt: { type: [string, 'null'], format: date-time }
        lastCheckStatus: { type: [string, 'null'] }
        source: { type: string, enum: [managed, environment] }
    CreateProviderRequest:
      type: object
      required: [name, base_url, api_key]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        base_url: { type: string, format: uri, description: HTTPS API root; local HTTP is allowed only for loopback testing }
        default_model: { type: string, maxLength: 128 }
        api_key: { type: string, minLength: 1, writeOnly: true }
        enabled: { type: boolean, default: true }
      additionalProperties: false
    UpdateProviderRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        base_url: { type: string, format: uri }
        default_model: { type: string, maxLength: 128 }
        api_key: { type: string, minLength: 1, writeOnly: true }
        clear_api_key: { type: boolean }
        enabled: { type: boolean }
      additionalProperties: false
...